backend/grapes_api/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
bleach==6.2.0
celery==5.5.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
django-redis==5.4.0
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
mysqlclient==2.2.7
packaging==25.0
pi==0.1.2
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-whois==0.9.5
PyYAML==6.0.2
redis==6.0.0
referencing==0.36.2
rpds-py==0.25.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
